requirements.txt
numpy==1.19.4
matplotlib==3.3.3
adafruit-circuitpython-lsm9ds1==2.1.3
influxdb==5.3.1
influxdb-client
pandas
scipy
scikit-learn
sympy

cycler==0.10.0
dtw-python==1.1.10
fastdtw==0.3.4
kiwisolver==1.3.1
Pillow==8.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0

sklearn